backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy[asyncio]
aiopg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cryptography
httpx
python-dotenv
pandas
aiofiles
celery
redis
flower
python-docx
beautifulsoup4
google-api-python-client
google-auth-oauthlib
lxml
pillow
shortuuid
boto3
ffmpeg-python
